# Perf: ReasonPhrases as Array instead of Dictionary