# ReasonPhrases as Array instead of Dictionary

- [x] You've read the [Contributor Guide](https://github.com/dotnet/aspnetcore/blob/main/CONTRIBUTING.md) and [Code of Conduct](https://github.com/dotnet/aspnetcore/blob/main/CODE-OF-CONDUCT.md).
- [x] You've included unit or integration tests for your change, where applicable.
- [x] You've included inline docs for your change, where applicable.
- [ ] There's an open issue for the PR that you are making. If you'd like to propose a new feature or change, please open an issue to discuss the change or find an existing issue.

Summary of the changes (Less than 80 chars)
Perf improvement ReasonPhrase

## Description
Replaced the Dictionary based implementation for an Array.

Benchmarks show an overall speed-improvement:
```
BenchmarkDotNet v0.13.12, Windows 11 (10.0.22631.3737/23H2/2023Update/SunValley3)
AMD Ryzen 5 2600X, 1 CPU, 12 logical and 6 physical cores
.NET SDK 9.0.100-preview.5.24307.3
  [Host]     : .NET 9.0.0 (9.0.24.30607), X64 RyuJIT AVX2
  DefaultJob : .NET 9.0.0 (9.0.24.30607), X64 RyuJIT AVX2
```

`UnchangedBenchmarks`

| Method           | Mean      | Error     | StdDev    | Ratio |
|----------------- |----------:|----------:|----------:|------:|
| Reasoning_Dict   | 42.563 ns | 0.2700 ns | 0.2255 ns |  1.00 |
| Reasoning_FrDict | 24.774 ns | 0.4166 ns | 0.3693 ns |  0.58 |
| Reasoning_Switch | 35.713 ns | 0.3052 ns | 0.2855 ns |  0.84 |
| Reasoning_Array  |  7.391 ns | 0.0836 ns | 0.0782 ns |  0.17 |

`PreparedBenchmarks`

| Method           | Mean      | Error     | StdDev    | Ratio |
|----------------- |----------:|----------:|----------:|------:|
| Reasoning_Dict   | 43.230 ns | 0.4669 ns | 0.4139 ns |  1.00 |
| Reasoning_FrDict | 23.568 ns | 0.3186 ns | 0.2980 ns |  0.55 |
| Reasoning_Switch | 36.178 ns | 0.4161 ns | 0.3689 ns |  0.84 |
| Reasoning_Array  |  7.674 ns | 0.0841 ns | 0.0703 ns |  0.18 |

`UnpreparedBenchmarks`

| Method           | Mean      | Error    | StdDev   | Ratio |
|----------------- |----------:|---------:|---------:|------:|
| Default          | 144.68 ns | 0.828 ns | 0.775 ns |  1.00 |
| FrozenDictionary | 132.87 ns | 1.394 ns | 1.304 ns |  0.92 |
| Switch           | 205.88 ns | 0.979 ns | 0.868 ns |  1.42 |
| Reasoning_Array  |  89.85 ns | 0.773 ns | 0.723 ns |  0.62 |

`StraightBenchmarks`

| Method           | StatusCode | Mean      | Error     | StdDev    | Median    | Ratio | RatioSD |
|----------------- |----------- |----------:|----------:|----------:|----------:|------:|--------:|
| Reasoning_Dict   | 0          | 3.4723 ns | 0.0506 ns | 0.0449 ns | 3.4748 ns |  1.00 |    0.00 |
| Reasongin_FrDict | 0          | 1.6297 ns | 0.0185 ns | 0.0173 ns | 1.6282 ns |  0.47 |    0.01 |
| Reasoning_Switch | 0          | 1.8411 ns | 0.0337 ns | 0.0315 ns | 1.8298 ns |  0.53 |    0.01 |
| Reasoning_Array  | 0          | 0.5079 ns | 0.0420 ns | 0.0392 ns | 0.4980 ns |  0.15 |    0.01 |
|                  |            |           |           |           |           |       |         |
| Reasoning_Dict   | 100        | 3.7815 ns | 0.0619 ns | 0.0579 ns | 3.7654 ns |  1.00 |    0.00 |
| Reasongin_FrDict | 100        | 1.7862 ns | 0.0479 ns | 0.0448 ns | 1.8067 ns |  0.47 |    0.02 |
| Reasoning_Switch | 100        | 0.7871 ns | 0.0166 ns | 0.0130 ns | 0.7872 ns |  0.21 |    0.00 |
| Reasoning_Array  | 100        | 0.8080 ns | 0.0158 ns | 0.0140 ns | 0.8076 ns |  0.21 |    0.01 |
|                  |            |           |           |           |           |       |         |
| Reasoning_Dict   | 102        | 3.7176 ns | 0.0398 ns | 0.0333 ns | 3.7167 ns |  1.00 |    0.00 |
| Reasongin_FrDict | 102        | 1.7751 ns | 0.0270 ns | 0.0226 ns | 1.7662 ns |  0.48 |    0.01 |
| Reasoning_Switch | 102        | 0.7865 ns | 0.0056 ns | 0.0053 ns | 0.7865 ns |  0.21 |    0.00 |
| Reasoning_Array  | 102        | 0.8378 ns | 0.0332 ns | 0.0310 ns | 0.8271 ns |  0.23 |    0.01 |
|                  |            |           |           |           |           |       |         |
| Reasoning_Dict   | 200        | 3.7126 ns | 0.0265 ns | 0.0207 ns | 3.7176 ns |  1.00 |    0.00 |
| Reasongin_FrDict | 200        | 1.7646 ns | 0.0142 ns | 0.0118 ns | 1.7649 ns |  0.47 |    0.00 |
| Reasoning_Switch | 200        | 1.5469 ns | 0.0085 ns | 0.0066 ns | 1.5469 ns |  0.42 |    0.00 |
| Reasoning_Array  | 200        | 0.8134 ns | 0.0233 ns | 0.0206 ns | 0.8102 ns |  0.22 |    0.01 |
|                  |            |           |           |           |           |       |         |
| Reasoning_Dict   | 226        | 4.7948 ns | 0.1199 ns | 0.1121 ns | 4.7392 ns |  1.00 |    0.00 |
| Reasongin_FrDict | 226        | 1.7394 ns | 0.0296 ns | 0.0263 ns | 1.7299 ns |  0.36 |    0.01 |
| Reasoning_Switch | 226        | 1.8112 ns | 0.0230 ns | 0.0192 ns | 1.8049 ns |  0.38 |    0.01 |
| Reasoning_Array  | 226        | 0.8106 ns | 0.0238 ns | 0.0223 ns | 0.8013 ns |  0.17 |    0.01 |
|                  |            |           |           |           |           |       |         |
| Reasoning_Dict   | 300        | 3.8204 ns | 0.1160 ns | 0.1548 ns | 3.8146 ns |  1.00 |    0.00 |
| Reasongin_FrDict | 300        | 2.2715 ns | 0.0330 ns | 0.0293 ns | 2.2635 ns |  0.59 |    0.03 |
| Reasoning_Switch | 300        | 0.7976 ns | 0.0246 ns | 0.0230 ns | 0.7943 ns |  0.21 |    0.01 |
| Reasoning_Array  | 300        | 0.8120 ns | 0.0344 ns | 0.0322 ns | 0.7936 ns |  0.21 |    0.01 |
|                  |            |           |           |           |           |       |         |
| Reasoning_Dict   | 308        | 3.7417 ns | 0.0167 ns | 0.0130 ns | 3.7447 ns |  1.00 |    0.00 |
| Reasongin_FrDict | 308        | 1.7781 ns | 0.0235 ns | 0.0196 ns | 1.7719 ns |  0.48 |    0.01 |
| Reasoning_Switch | 308        | 0.8056 ns | 0.0068 ns | 0.0060 ns | 0.8064 ns |  0.22 |    0.00 |
| Reasoning_Array  | 308        | 0.8743 ns | 0.0727 ns | 0.0866 ns | 0.8093 ns |  0.23 |    0.02 |
|                  |            |           |           |           |           |       |         |
| Reasoning_Dict   | 499        | 3.7423 ns | 0.0429 ns | 0.0359 ns | 3.7348 ns |  1.00 |    0.00 |
| Reasongin_FrDict | 499        | 1.7812 ns | 0.0426 ns | 0.0378 ns | 1.7664 ns |  0.48 |    0.01 |
| Reasoning_Switch | 499        | 2.0517 ns | 0.0138 ns | 0.0129 ns | 2.0481 ns |  0.55 |    0.01 |
| Reasoning_Array  | 499        | 0.8138 ns | 0.0283 ns | 0.0251 ns | 0.8177 ns |  0.22 |    0.01 |
|                  |            |           |           |           |           |       |         |
| Reasoning_Dict   | 500        | 3.7223 ns | 0.0179 ns | 0.0149 ns | 3.7173 ns |  1.00 |    0.00 |
| Reasongin_FrDict | 500        | 1.7743 ns | 0.0316 ns | 0.0296 ns | 1.7621 ns |  0.48 |    0.01 |
| Reasoning_Switch | 500        | 1.6978 ns | 0.0165 ns | 0.0146 ns | 1.6946 ns |  0.46 |    0.00 |
| Reasoning_Array  | 500        | 0.8439 ns | 0.0604 ns | 0.0565 ns | 0.8200 ns |  0.23 |    0.02 |
|                  |            |           |           |           |           |       |         |
| Reasoning_Dict   | 511        | 3.8137 ns | 0.0504 ns | 0.0447 ns | 3.7935 ns |  1.00 |    0.00 |
| Reasongin_FrDict | 511        | 1.7733 ns | 0.0152 ns | 0.0142 ns | 1.7725 ns |  0.47 |    0.01 |
| Reasoning_Switch | 511        | 2.0656 ns | 0.0061 ns | 0.0057 ns | 2.0656 ns |  0.54 |    0.01 |
| Reasoning_Array  | 511        | 0.8160 ns | 0.0138 ns | 0.0122 ns | 0.8159 ns |  0.21 |    0.00 |

Edit: updated benchmarks to show results according to https://github.com/dotnet/aspnetcore/pull/56304#issuecomment-2183338237
Edit2: updated benchmarks to show results according to https://github.com/dotnet/aspnetcore/pull/56304#issuecomment-2190641197
Edit3: updated benchmarks to show results according to https://github.com/dotnet/aspnetcore/pull/56304/commits/344293dfe2500a3a568c97a3a323b4acab314ae3