@gfoidl 2024-06-20
@WhatzGames can you please show the benchmark code?

If the benchmarks have predictable input, then the CPU's branch predictor will do a good job (i.e. it predicts which branch to take) so that can bias the results.

It would be interesting to see these improvements w/ random status codes too.
Maybe instead of the Dictionary a FrozenDictionary should be tried too.